4-5-26
... update status

+++ database/migrations/2026_05_04_045151_add_parent_id_to_menus_table.php
@@ -6,25 +6,22 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
+
     public function up(): void
     {
         Schema::table('menus', function (Blueprint $table) {
-            // Menambahkan kolom parent_id yang merelasikan ke tabel menus itu sendiri
-            // Posisinya diletakkan setelah kolom 'id' agar rapi di database
-            $table->foreignId('parent_id')
-                  ->nullable()
-                  ->constrained('menus')
-                  ->onDelete('cascade')
-                  ->after('id'); 
+            if (!Schema::hasColumn('menus', 'parent_id')) {
+                // Menambahkan kolom parent_id yang merelasikan ke tabel menus itu sendiri
+                // Posisinya diletakkan setelah kolom 'id' agar rapi di database
+                $table->foreignId('parent_id')
+                      ->nullable()
+                      ->constrained('menus')
+                      ->onDelete('cascade')
+                      ->after('id'); 
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('menus', function (Blueprint $table) {
